+++ frontend/src/types/preferences.ts
@@ -4,3 +4,7 @@ export interface Preference {
     name: PreferecnceName;
     value: string;
 }
+
+export interface Preferences {
+    preferences: Preference[];
+}