+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Rent status is now displayed ([@kovaj024](https://github.com/kovaj024))
 - Tenant overview page ([@kovaj024](https://github.com/kovaj024))
 - Marking rent status via UI ([@kovaj024](https://github.com/kovaj024))
+- Settings page to set preferences ([@zvonimirr](https://github.com/zvonimirr))
 
 ## 0.1.0